Moved GameDetail component into its own page
